tools/tralfaz-telegram/requirements.txt
python-telegram-bot
anthropic
httpx
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
